Implement general global load/store.
Supports single globals and trees transparently.

Includes a large move of classes in order to break circular dependencies. Will re-organize later.
