+++ python/shark_turbine/aot/procedural.py
@@ -5,139 +5,43 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-from contextlib import contextmanager
 import re
-import threading
 from typing import Any, Callable, List, Sequence
 
 import torch
-
-from iree.compiler.ir import (
-    FunctionType,
-    InsertionPoint,
-    Location,
-    RankedTensorType,
-    Type as IrType,
-    TypeAttr,
-    Value,
-)
 
 from iree.compiler.dialects import (
     func as func_d,
 )
 
-from torch.utils._pytree import (
-    tree_map,
+from numpy import number
+
+from ..dynamo.importer import TORCH_DTYPE_TO_MLIR_TYPE_ASM
+
+from .support import (
+    AbstractIntrinsic,
+    Intrinsic,
+    IrTrace,
+    LiveGlobalCollectionProxy,
+    Location,
+    ModuleBuilder,
+    StringAttr,
+    Value,
+    logger,
+    new_ir_trace_scope,
     tree_flatten,
     tree_unflatten,
+    treespec_dumps,
 )
-
-from numpy import number
-from collections.abc import Mapping
-
-from .builder import ModuleBuilder
-from ..dynamo.importer import TORCH_DTYPE_TO_MLIR_TYPE_ASM
 
 # We need the inverse of the TORCH_DTYPE_TO_MLIR_TYPE_ASM table.
 MLIR_TYPE_ASM_TO_TORCH_DTYPE = {v: k for k, v in TORCH_DTYPE_TO_MLIR_TYPE_ASM.items()}
-
-_thread_state = threading.local()
-
-
-class Intrinsic:
-    """Objects which interact natively with the tracing system implement this."""
-
-    __slots__ = []
-
-    def resolve_ir_values(self, proc_trace: "ProcedureTrace") -> Sequence[Value]:
-        raise NotImplementedError(
-            f"Cannot use {self} as an expression in a procedural function"
-        )
-
-    def resolve_call(self, proc_trace: "ProcedureTrace", *args, **kwargs):
-        raise NotImplementedError(
-            f"Cannot use {self} as the target of a call in a procedural function"
-        )
-
-
-class CallableIntrinsic(Intrinsic):
-    """Intrinsic subclass that supports calls.
-
-    This is separate so as to make error handling better (i.e. does not support
-    calls) for intrinsics that are not callable.
-    """
-
-    __slots__ = []
-
-    def __call__(self, *args, **kwargs):
-        return current_ir_trace().handle_call(self, args, kwargs)
-
-
-class AbstractIntrinsic:
-    """Base class for descriptor types that can be converted to Python proxies."""
-
-    __slots__ = []
-
-    def create_intrinsic(self, value: Value) -> Intrinsic:
-        """Creates a proxy object that can flow through a procedural trace."""
-        raise NotImplementedError
-
-    def get_ir_type(self, builder: ModuleBuilder) -> IrType:
-        """Gets the corresponding IR type."""
-        raise NotImplementedError
-
-
-class IrTrace:
-    """Gets callbacks for tracing events."""
-
-    __slots__ = []
-
-    def finalize(self):
-        """Called when the trace is finished (popped off the stack)."""
-        pass
-
-    def handle_call(self, target: Intrinsic, args, kwargs):
-        raise NotImplementedError(f"The current trace scope does not support calls")
-
-
-class ImmediateIrTrace(IrTrace):
-    __slots__ = []
-    ...
-
-
-def _trace_scopes() -> List[IrTrace]:
-    try:
-        trace_scopes = _thread_state.trace_scopes
-    except AttributeError:
-        trace_scopes = _thread_state.trace_scopes = [ImmediateIrTrace()]
-    return trace_scopes
-
-
-@contextmanager
-def new_ir_trace_scope(ir_trace: IrTrace):
-    trace_scopes = _trace_scopes()
-    trace_scopes.append(ir_trace)
-    try:
-        yield ir_trace
-    finally:
-        ir_trace.finalize()
-        del trace_scopes[-1]
-
-
-def current_ir_trace() -> IrTrace:
-    return _trace_scopes()[-1]
 
 
 class ProcedureTrace(IrTrace):
     """Captures execution of a Python func into IR."""
 
     __slots__ = [
-        "module_builder",
-        "func_op",
-        "context",
-        "ip",
-        "return_types",
-        "loc",
         "proxy_posargs",
         "proxy_kwargs",
     ]
@@ -150,12 +54,7 @@ class ProcedureTrace(IrTrace):
         proxy_posargs,
         proxy_kwargs,
     ):
-        self.module_builder = module_builder
-        self.func_op = func_op
-        self.context = func_op.context
-        self.ip = InsertionPoint(self.func_op.entry_block)
-        self.return_types = None
-        self.loc = self.func_op.location
+        super().__init__(module_builder=module_builder, func_op=func_op)
         self.proxy_posargs = proxy_posargs
         self.proxy_kwargs = proxy_kwargs
 
@@ -191,6 +90,12 @@ class ProcedureTrace(IrTrace):
             ir_proxy_arguments_flat, arguments_tree_def
         )
 
+        # Metadata.
+        if arguments_flat:
+            func_op.attributes["torch.args_schema"] = StringAttr.get(
+                treespec_dumps(arguments_tree_def), context=module_builder.context
+            )
+
         return ProcedureTrace(
             module_builder=module_builder,
             func_op=func_op,
@@ -201,38 +106,83 @@ class ProcedureTrace(IrTrace):
     def trace_py_func(self, py_f: Callable):
         with new_ir_trace_scope(self) as t:
             # TODO: Create IR proxies for python arguments.
-            return_py_value = py_f(*self.proxy_posargs, **self.proxy_kwargs)
+            return_py_value = _unproxy(py_f(*self.proxy_posargs, **self.proxy_kwargs))
             if return_py_value is None:
                 self.emit_return()
             else:
-                flat_return_py_values, _ = tree_flatten(return_py_value)
+                flat_return_py_values, schema = tree_flatten(return_py_value)
                 flat_return_ir_values = []
                 for py_value in flat_return_py_values:
                     flat_return_ir_values.extend(convert_py_value_to_ir(self, py_value))
+                self.func_op.attributes["torch.return_schema"] = StringAttr.get(
+                    treespec_dumps(schema), context=self.context
+                )
                 self.emit_return(*flat_return_ir_values)
-
-    def emit_return(self, *ir_values: Sequence[Value]):
-        with self.loc, self.ip:
-            func_d.ReturnOp(ir_values)
-            # Check or rewrite the function return type.
-            value_types = [v.type for v in ir_values]
-            if self.return_types:
-                if value_types != self.return_types:
-                    raise ValueError(
-                        f"Multi-return function must return same types. "
-                        f"{value_types} vs {self.return_types}"
-                    )
-                return
-            self.return_types = value_types
-            ftype = self.func_op.type
-            ftype = FunctionType.get(ftype.inputs, value_types)
-            self.func_op.attributes["function_type"] = TypeAttr.get(ftype)
-            assert self.func_op.verify(), "Created function is invalid"
 
     def handle_call(self, target: Intrinsic, args, kwargs):
         """Implements calls to jittable functions."""
         with self.loc, self.ip:
             return target.resolve_call(self, *args, **kwargs)
+
+    def handle_assignment(self, scope, target, updated_value):
+        logger.debug(
+            "ASSIGN %r.%r = %r", scope.__class__, target.__class__, updated_value
+        )
+        self._recursive_assign(target, updated_value, set())
+
+    def _recursive_assign(self, target, source, encountered_set):
+        target = _unproxy(target)
+        source = _unproxy(source)
+
+        # Check for cycles.
+        target_id = id(target)
+        if target_id in encountered_set:
+            raise TypeError(f"Cycle in tree assignment target")
+        encountered_set.add(target_id)
+
+        # Leaves/terminals.
+        if isinstance(target, Intrinsic):
+            if not isinstance(source, Intrinsic):
+                raise TypeError(
+                    f"Cannot assign mismatched leaf types in a tree: "
+                    f"{target.__class__} vs {source.__class__}"
+                )
+            leaf_values = source.resolve_ir_values(self)
+            target.resolve_assignment(self, leaf_values)
+            return
+
+        # Zip across dicts.
+        if isinstance(target, dict):
+            if not isinstance(source, dict):
+                raise TypeError(
+                    f"Mismatched dict assignment in a tree: {target.__class__} vs {source.__class__}"
+                )
+            target_keys = target.keys()
+            source_keys = source.keys()
+            if target_keys != source_keys:
+                raise TypeError(
+                    f"Mismatched dict keys in tree assignment: {target_keys} vs {source_keys}"
+                )
+            for k in target_keys:
+                target_child = target[k]
+                source_child = source[k]
+                self._recursive_assign(target_child, source_child, encountered_set)
+            return
+
+        # Zip across lists/tuples (we let them be used interchangeably at the source).
+        if isinstance(target, list):
+            if not isinstance(source, (list, tuple)):
+                if len(target) != len(source):
+                    raise TypeError(
+                        f"Mismatched sequence length in tree assignment: {len(target)} vs {len(source)}"
+                    )
+            for target_child, source_child in zip(target, source):
+                self._recursive_assign(target_child, source_child, encountered_set)
+            return
+
+        raise TypeError(
+            f"Cannot recursively assign through a container of {target.__class__}"
+        )
 
 
 class ProcedureTraceError(Exception):
@@ -252,59 +202,7 @@ def convert_py_value_to_ir(
     )
 
 
-class IrTensorBase(Intrinsic):
-    """Base class for 'tensors' that resolve to some IR value.
-
-    These are not real tensors (i.e. no operations can be performed on them),
-    but they stand in as reasonable proxies during procedural tracing.
-    """
-
-    __slots__ = [
-        "ir_type",
-        "dtype",
-    ]
-
-    def __init__(self, ir_type: IrType, dtype: torch.dtype):
-        self.ir_type = ir_type
-        self.dtype = dtype
-
-    @classmethod
-    def __torch_function__(cls, func, types, args=(), kwargs=None):
-        return NotImplemented
-
-    def _to_meta_tensor(self) -> torch.Tensor:
-        """Converts to a fake Tensor that dynamo can handle."""
-        ir_tensor_type = RankedTensorType(self.ir_type)
-        shape = ir_tensor_type.shape
-        # TODO: Remove this assert. We need to extend this method to also be
-        # able to contribute dynamic_dim constraints and then return a minimum
-        # quantity (2) for any dynamic dim.
-        assert not any(
-            d < 0 for d in shape
-        ), "Dynamic dims to jittable not yet implemented"
-
-        # TODO: We shouldn't need to create a real tensor here, as Dynamo will
-        # immediately convert it to fake. However, it will also set up the shape
-        # environment and asserts that any fake tensor inputs are from its
-        # internal FakeMode. There should be a way but needs more investigation.
-        # TODO: This tensor needs a device that matches the model being exported.
-        # We just create these on the CPU because that is common.
-        return torch.empty(shape, dtype=self.dtype)
-
-
-class IrValueTensor(IrTensorBase):
-    """Represents a Value in the IR under construction during procedural tracing."""
-
-    __slots__ = [
-        "ir_value",
-    ]
-
-    def __init__(self, ir_value: Value, dtype: torch.dtype):
-        super().__init__(ir_value.type, dtype)
-        self.ir_value = ir_value
-
-    def __repr__(self):
-        return f"IrValueTensor(@{self.ir_value})"
-
-    def resolve_ir_values(self, proc_trace: ProcedureTrace) -> Sequence[Value]:
-        return (self.ir_value,)
+def _unproxy(value):
+    if isinstance(value, LiveGlobalCollectionProxy):
+        return value._raw_collection
+    return value
